Improved display, cleaned up more hard-coded dimensions.
The display is more square now. This depends on your terminal font, of
course, but for most it should be a lot more pleasing.

The display also doesn't assume that the board is 8x8 anymore. This
should make it much easier to convert this into another game such
a Draughts, which is played on a 10x10 board.